refactor: impliment psr4 for Reports
