+++ give.php
@@ -595,7 +595,7 @@ if ( ! class_exists( 'Give' ) ) :
 			require_once GIVE_PLUGIN_DIR . 'blocks/load.php';
 
 			//Include new reports page
-			require_once GIVE_PLUGIN_DIR . 'includes/reports/class-reports.php';
+			require_once GIVE_PLUGIN_DIR . 'src/reports/Reports.php';
 
 			if ( defined( 'WP_CLI' ) && WP_CLI ) {
 				require_once GIVE_PLUGIN_DIR . 'includes/class-give-cli-commands.php';
