+++ src/Reports/Report/Payments.php
@@ -5,15 +5,13 @@
  * @package Give
  */
 
-namespace Give;
+namespace Give\Reports\Report;
 
 defined( 'ABSPATH' ) || exit;
-
-require_once GIVE_PLUGIN_DIR . 'includes/reports/reports/class-report.php';
 
 /**
  * Functionality and actions specific to the payments report
  */
-class Payments_Report extends Report {
+class Payments extends Report {
 
 }
